Redesign task scheduling setup
This is a major redesign of the scheduling and task switching setup. The
main goal of this design is to redo task-switching to take place between
actual user-threads vs. switching between kernel-threads for each
user-thread. This allows us to only use one kernel stack per-cpu,
instead of one kernel-stack per user-thread.

The new design works by saving the context into the struct task *, and
then on an interrupt saving the old iframe data and replacing it with a
new task's information. That way, when we return from an interrupt after
a scheduling, we return into the new task instead of to the old one.

As like before, because the filesystem is not yet in a workable state
the running tasks are manually created.