Redid some tasking and interrupt stuff
Now instead of simply deadlocking when there are no runnable processes,
instead a per-cpu 'idle' kernel task is switched to. The task is
interruptable so we can still handle interrupts in that time. Also
implemented, a separation between 'interrupt' irq's and 'syscall'
irq's. The difference indicates what you can do in the interrupt
context. syscalls allow you to suspend the currently running task with
'task_yield' or 'task_sleep'. Interrupt context's are not allowed to do
that. Instead, interrupt contexts can set the 'reschedule' flag to
indicate a new process shoudld be scheduled once the interrupt returns.
The timer interrupt uses the 'reshedule' flag.

Also fixed, eflags is now pushed by arch_context_switch. This previously
wasn't pushed, but that was an error since that state does need to be
saved. This issue is more predominant with interruptable and
noninterruptable kernel tasks, which require backing up the eflags to
back-up whether or not interrupts were enabled or disabled for this
kernel task.