More big changes
The biggest change here (besides kp()) is probably the move from
simple_fs, to ext2. The ext2 driver is currently read-only, but works to
the point that we can execute programs. This is the same functionality
that simple_fs can currently do, but ext2 supports much more then
simple_fs can.

Note that this change nesisitates a different way to build the
./disk.img. It's necessary to use sudo in the Makefile to get root to
mount the ext2 image.