Lots of changes
Implement user-pointer checking - Pointers to user-memory that are
passed via syscalls to the kernel are now checked to ensure they point
to valid pieces of memory.

Also implemented is a proper COM driver. Along with this, the
com1_debug.c driver was removed, and the kprintf system was reworked to
instead have a 'register' system for registering outputs from kprintf.
On boot, the console and com1 are registered, after booting the
console is removed. This system can be improved upon (And kprintf really
needs a full log stored in the kernel), but that opens up other problems
and this is sufficent for now. With this change, the /dev/com1 and
/dev/com2 devices are now in /dev.