Fix subtle bug in initial task setup
It turns out that compound literals in C do not have a guarenteed
life-time past the exit of the function even if they're read-only. Thus,
the contents of the compound literal used to initialze the argv and envp
is actually placed on the stack and thus randomly overwritten. The
simple fix is to instead just pass NULL for both, since sys_execve
handles NULL perfectly fine.