Fix incorrect types due to typedef
`const va_t` and `const void *` are not the same thing. `const va_t`
means `void *const`. Unfortunately there is no way to express
`const void *` using `va_t`, so instead the declaration is replaced with
`const void *`.

Generally speaking, the distinction between the const and non-const
versions of `va_t` doesn't actually matter, because va_t is not meant to
be dereferenced on its own except in special cases where it is converted
into another type of pointer. Nevertheless, it does make sense to make
the argument to user_check_region `const` to avoid warnings. This is
only possible by explicetly requiring a `const void *` instead of
`va_t`.