Prevent page-fault loop if backtrace faults
When the kernel page-faults, the handler dumps the state of the kernel
and exits. In this dump, it attempts to print a backtrace of the
functions called both by the kernel, and by the running user process (if
anyway). There's however, a possibility that the stack no longer points
at valid memory. If this happens, the backtrace would page-fault,
leading to the handler being run again, the backtrace being done again,
leading to another page-fault, etc.

To keep this from happening, a static variable was added into the
page-fault handler, and is marked when a backtrace is about to be done.
If the backtrace page-faults, the second entry to the handler will check
the static variable and then skip the backtrace when it is set.