Fix bug, and add bit_find_next_zero function
The function had both 'i' and 'index' variables. When returning the
index, the wrong one was used, so the wrong value was returned. As the
function was never really used the bug was not noticed.

In addition, the function bit_find_next_zero was added, which takes an
offset indicating which bit to start scaning from.