Remove TASK_RUNNABLE state
TASK_RUNNING and TASK_RUNNABLE complicated the scheduling and are
unnecessary. The TASK_RUNNABLE state was removed, and a new 'running'
flag was added to struct task. The 'running' flag indicates where or not
a task is currently being run on a processor, which is what the
distinction between TASK_RUNNABLE and TASK_RUNNING was.