Fix ext2 file truncate
The bug was due to incorrect usage of the 'starting_block' and
'ending_block' markers used to indicate where to start and stop
truncating blocks. 'starting_block' is rounded up, 'ending_block' is
rounded down. When truncating, you delete both 'starting_block' and
'ending_block', not 'up to' ending_block.

In addition, the 1 subtraction to the block number should happen before
calculating index and group number.