Fix kernel exported types setup
This change both fixes the declarations of some types, and also creates
the __k prefix'd types. These prefix'd types are exported by the kernel
to userspace so they don't conflict with the normal types (And guarentee
the correct width). Along with that, the libc can be modified to use
these type declarations to guarentee the correct types are used.