Use kwork for tty, and add setuid
This change removes the `tty_kernel_task` that was previously used to
process input/output through the `tty` layer. Now, the `struct tty`
contains a `struct work` entry, and when input or output needs to be
processed through the `tty` layer, a kwork task is kicked-off to do it.
The 'register_for_wakeups' methods were then modified to register the
`struct work` instead of the task.

`struct work` was also changed to accept a regular `struct task *` entry
in lue of running an actual `struct work`. If `kwork_schedule` is called
on one of these `struct work` objects, the task is simply woken-up
instead.  This is to facilitate removing the `struct wakeup_list` usage,
as now threads can use the kwork system to wake for events along with
actual `struct work` entries.

This also adds the `setuid` syscall, mostly for testing.