Add userdel and usermod coreutils
This adds the `userdel` and `usermod` coreutil programs. The designed is
based on a backing read/write "db", which presumes a format of lines with `:`
separated entries. That's then converted into a more special format for
files like `/etc/passwd`. Another db for `/etc/group` will be added
next. `init` will also be moved to the coreutils and make use of the
`db` for `inittab`.