Mark __using_cond as const so gcc doesn't complain about return paths
When using the scoped_using like this:

    scoped_using()
        return blah;

This is perfectly valid, the cleanup method regsitered by
`scoped_using()` will be called when the `return` is executed. However
`gcc` can't quite prove that the `scoped_using()` will always execute
the body when teh `__using_cond` is not `const`, leading to a warning
about a possible missing return.