Redesign memory-mapping subsystem
This restructures a lot of the memory-management related to the `vm_map`
structures. The biggest change is the removal of the `page_list` from
the `vm_map`, with the replacement mapping pages directly into the page
directory.

This is in preparation for full `mmap` support and demand paging, where
the page-handler will back-end into the related `vm_map`, and then the
`vm_map` will support a few callbacks for handling loading in pages.