Refactor wait_queue to use struct work and not be exclusive
This patch refactors how wait_queues work to simplify the overall
implementation.

First, they now use a `struct work` to represent what to do on a wakeup.
The `struct work` can represent a variety of different use-cases and I'm
moving to use it in most cases in the kernel to unify
waiting/completion.

Second, the implementation is now not an exclusive `wait_queue`, but
instead simply wakes up everyone on the list. This simplifies the logic
greatly by not requiring the weird 'chaining' of wakeups to ensure no
wake-ups were lost. Waking up all tasks instead is much simpler to
ensure no wakeups are lost (at the cost of being less efficent, though
for the time being this isn't that big of a deal).