Move back to 1024 block size, bump size to 256MB
Unfortunately mkfs.ext2 still creates the directory index even when
requsted not too when using a 4096 block size. I'm not sure why, but I'm
going back to 1024 for the moment since we don't support the directory
index.

The size was also bumped to 256MB to prepare for more extras that take
up more space.