Disable db-install when building ncurses
Ubuntu Bionic's `tic` exe has trouble parsing the db, and obviously the
cross-compiled version can't be used. Not installing the DB is not ideal
but it does still mroe or less work.

The ultimate solution is to compile `ncurses` from within `protura`
rather than via a cross-compiler, but that would seriously complicate
the automated build.