ncurses: Refactor to build a host-version of tic
Using the built-in DB is just not sufficent for ncurses. So to fix this,
we're now building a host version of the ncurses progs, and then uses
those progs when building the cross-compiled version. It's a bit
convoluted, but it works fairly well and doesn't depend on the host
having the right version of ncurses.