Remove unnecessary static init_task
It turns out this was actually unecessary, once we have the pic8295
timer running, we're able to create a task as long as we ensure the
allocation does not sleep. To achieve this `task_kernel_generic` was
exposed, which allows us to allocate the task object ourself and ensure
the `PAL_ATOMIC` flag is used. From there, we just add the task to the
scheudler and then start the scheduler. All the rest of the init
tasks are then called from that kernel task.