ext2: Fix bugged block group calculation
Currently, the calculation of the total number of block groups assumes
that the total number of blocks in the disk evenly divides by the number
of blocks per group. But actually, If you have an oddly sized partition
these don't have to evenly divide, and we have to make sure we account
for the last possibly undersized block group at the end.