Block: Add partition support
This patch adds partition support, via modifying `block_device` to hold
multiple `partition` objects that represent the start and length of that
device. Then the block_size was modified, with it now being acquired via
a function, which checks the minor number and then returns the
block_size of the partition if the minor is greater than 0. And then the
IDE driver checks the minor number and adds the partition sector offset.

With this, the scripts were updated to generate a partitioned disk,
rather then a raw EXT2 disk. Along with that, GRUB is now installed on
the disk and the disk can be booted on its own. We also now offer a
`make` recipe to generate a VirtualBox disk image, which we will build
on Travis and publish with releases.