wait.h: Explicitly set the queue to NULL on unregsiter
This helps us catch some issues with wait_queues. It unfortunately also
complicates the code slightly as we read node->queue from outside the
queue lock (obviously), so a barrier is required to ensure the ordering.

There is an issue when a `wait_queue` is removed, but at the moment all
the wait_queues are static, so that's not currently an issue.