syscall.h: Fix syscall numbers
It turns out there were a few duplicates here, and there were also some
sections of numbers not used. This unifies all the numbers to ensure
they are all unique and incremented by one from eachother.

NOTE: This commit requires a full recompile of userspace to use the new
numbers.