Add ktest kernel testing framework
ktest is pretty simple and based off of some of my other C unit testing
frameworks. it is based off of two structures, the `struct ktest_module`
and `struct ktest_unit`. Every module has an array of `struct
ktest_unit` entrise, which are called in series. The `struct ktest_unit`
has a callback to run that set of tests, and then returns the number of
failures. All the failures are collected together and reported at the
end.

It also includes some convience macros in the form of `ktest_assert_*`
macros, which use _Geenric to detect the type of the input and do the
correct comparison.

We make use of a new `.ktest` section in the binary to find all of our
`struct ktest_module` entries at runtime. The macro __ktest can be used
to place a module into that section.

Testing can be turned on or off via the `CONFIG_KERNEL_TESTS` config
option.