slab: fix memory leak in __slab_frame_object_free
Because the loop stopped when *current was NULL (Even though the inner
part correctly handled this situation), when the list was empty or the
object being freed was at the end, the memory would be leaked instead of
returned to the freelist for that frame. The solution is to not check
for `*current == NULL` so this case is correctly handled by the inner
part of the loop.