ktest: Use ksetjmp, and remove returning int
We're now using ksetjmp and klongjmp to exit tests when an assertion
fails. This keeps us from continuing the test code which may cause the
kernel to crash due to the assertion failing.

We also removed the need to track and return the number of failed tests,
by instead keeping that count in the `struct ktest` structure (Though,
with the introduction of `ksetjmp` we won't continue running tests after
the first).