ktest: new argument system
This is a complete overhaul of the previous argument system for ktest.
It consisted of a single `unit->arg` value which could be set. Now,
KTEST_UNIT_INIT() is a varadic macro which can takeu up to 6 arguments
wrapped in KT_* macros. The arguments can then be recovered via
`KT_GET_ARG(kt, index, type)`, which does type checking to ensure the
provided type matches the type being asked for, and then returns the
value.

Making this work requires fairly heavy usage of `_Generic()` and some
other macro blackmagic. Most of it is documented in the code.