tty: remove input_buf
input_buf is unnecessary, as it is only written too when processes call
tty_write (Via an open fd). In this context, we're free to simply call
`tty_process_output` directly and skip the buffer entirely.

This has the big advantage that it solves an existing problem where, due
to the decupled nature of that buffer, a process can exit before all of
the data they wrote to the tty has actually been displayed.