char-devices: remove console, keyboard, com, and screen
The console subsystem is about to get a bit of an overhaul, and these
devices (which shouldn't be used, you should go though the relevant TTY
instead) need to go as that interfaces are likely to change.

Note that this change preserves the char device numbers, so it is not
necessary to change those.