console: restructure, and add basic VT100 support
After the remove of term.c this is a large restructuring of the console
output. There is now a `vt` layer (which is mostly generic) which
implements various VT100 escape codes. After that, there is the `screen`
and `keyboard` layers which are arch specific, and then the `console`
layer which ties everything together into a tty device.