packet: remove socket reference
The reference to the socket that created a paticular packet was
unnecessary - the packat contains all of the necessary information. The
reason for removing it is to simplfy the upcomming close()/shutdown()
changes. The `struct socket` object lifetime is going to be
disassociated from the `struct file` used to interact with it.

By not having the socket and packet structures tied together, we
don't have to worry about some packets being inflight that still hold a
reference to this socket.