spinlock: fix save/restore of EFLAGS
We were mistakenly writing the `lock->eflags` value before we actually
acquire the lock. It's correct that we need to save the EFLAGS from
before we do cli() and spin for the lock, but it's incorrect to
immedieatly write that to the spinlock itself, since that overrides the
EFLAGS for the current owner, potentially completely locking things up.

The fix is to save the EFLAGS to a temporary, and then write that
temporary into the lock after it is acquired.