net: remove unnecessary socket_inode_delete
Because the socket reference held in the inode isn't actually duped and
the same one is held in the `struct file`, the socket_put() in here was
a bug. It's easy enough to just remove the delete - the anonymous inode
won't ever be associated with another `struct file`, to tying the socket
reference to the file is simpler.