socket: add socket_release
We're separating the concept of shutdown() (a graceful close requested
on the socket) vs a release() (When the socket no longer has a user).
This is important because shutdown() only happens on connected sockets
(And thus never does anything for UDP/connectionless sockets). release()
however closes down the socket and prepares it to be released, which is
necessary to avoid leaking the sockets after they are done being used.