conf: increase default task switches per second
This was previously set to 100, which was lay to low and the cause of a
lot of the speed issues related to disk speed .Even if the disk
responded almost instantinously, we were stuck waiting 10ms to get
rescheduled again, since the idle process would start.

We're increasing this to 2000 (along with increasing the max timer ticks to
match) so that we reschedule tasks much quicker and reduce the latency
when sleeping.