ext2: fix block allocation
We came across a bug when we switched from using a 1024 block size ext2
drive to a 4K ext2 block drive, where it would allocate the blocks
incorrectly. We 'fixed' this by removing a '+1' from the block
allocation, but that was really just hiding the real bug.

The ext2 super block stores the block number the super-block is stored
in. This value is zero for a 4K block size, but 1 for a 1024 block size.
It is this value that we need to add and subtract from the block numbers
when we allocate and release them.