ext2: fix i_blocks calculation
The 'blocks' value on ext2 inodes is supposed to count every block in
use for storing the data, including the indirect blocks. We were
incorrectly excluding those, leading to a bug.

To fix this, we provide the ext2_inode when doing the block allocation
or release, and do the modification of the 'blocks' value there.