ide_dma: correctly set the RWCON bit
This fixes IDE DMA on VirtualBox (When using the PIIX3 chipset)

The documentation on RWCON is somewhat misleading (to me, anyway). It
says that RWCON being set indicates a "write", and RWCON being cleared
indicates a "read". However, I believe this is written from the
perspective of the DMA controller writing to and reading from memory, so
from that perspective, when we're "writing" to the disk, the DMA
controller is "reading" from memory, and vice versa.

The result is that we need to set the bit when reading from the disk,
and clear the bit when writing to the disk. We also need to make sure we
set this bit before setting the bus mastering bit on (The SSBM bit).