packet: add packet_take_page()
packet_take_page() allows someone that owns the packet to take the
underlying page from the packet so it can be used independently. The
page is diassociated from the packet, so the packet's data can no longer
be read, and the page will no longer be free'd when packet_free() is
called, pfree() should be called manually instead.