tcp: enable sending RST packets without a socket
This restructures some of the TCP and IP output code so that we can send
a TCP packet without it being tied to a specific TCP socket. This is
important for RST packets as they're likely not associated with any
paticular active socket.

There is also a small change for the TCP FIN handling - the FIN packet
is added to the recv and out-of-order queue so that it is correctly
processed in order relative to the other data packets. This ensure we
don't close the socket when we have yet to recieve all of the data from
the other side. It's not extremely relevant since we still don't do
proper TCP close, but it is a step in the right direction.