keyboard: complete redo keyboard support
There were lots of things broken about our keyboard support, with the
biggest thing being the lack of support for any extended keys past the
basic ones. There was also the lack of any real generic layer.

This restructure takes a lot of inspiration from the Linux keyboard
subsystem, though it is a fair bit simplified. The architecture specific
code for the keyboard reads the scancodes and converts them into
`keysym` values. The keysym's are then fed into our generic keyboard
handling layer, which then acts on the `keysym` utilizing our `keymap`
array.

For the x86 keyboard driver, we now correctly handle E0 scancodes and
map them into new keysym values. With this, the bulk of the keyboard
logic was removed from this driver, and it simply feeds the ksysyms into
the generic driver.

For the generic driver, there are 8 different keymaps dependent on the
status of the 3 keyboard modifiers (shift, control, alt). The keymap
used is selected based on which of these modifiers is currently active.

The keymaps themselves are big arrays of `struct keycode` values. Each
value holds a keytype, indicate the type of action to happen when that
keysym is entered, and then an argument to that keycode whos value
depends on the keytype in use. For regular keycodes, that value is just
the ASCII value typed in. For others, there are other enumerations
indicating the valid values.

From there, when the selected keycode resolves to an actuall ASCII value
or values, those are entered into the currently active TTY like before.