seq_file: add overflowed flag
The overflowed flag is stored in the `struct seq_file` and marks if any
seq_printf operations on the object fail with -ENOSPC. If, after the
`render` callback is run, the overflowed flag is set, then the kbuf is
automatically reset to the position before `render` was called.

The purpose of this is to make the seq_file API automatically reset the
kbuf if any seq_printf's fail in the `render`. This allos the `render`
callback to ignore the results of any `seq_printf` and also not have to
manually handle rolling back if they do multiple `seq_printf`s in a row.