tty: remove tty_array and make device_no the actual device
This is somewhat simple but still substantial redesign of the tty
devices. The `tty_array` that consisted of every `struct tty` is gone,
instead they are allocated on demand when a `struct tty_driver` is
registered.

This change allows us to have tty's exist on different major numbers. To
make this work, `device_no` on the `struct tty` is changed to be the
actual device number for that `struct tty`, and the existing code is
changed to respect that.