ide: reading/writing sectors should wait for data-request status
This fixes writing to the disk on Hyper-V.

When reading or writing sectors to the IDE drive, we have to wait for
the 'data-request' status from the drive before doing a read/write of an
individual sector.

Previously, we were waiting for the 'drive-ready' status, but that is
actually incorrect. When the drive is ready to read/write a sector, it
asserts the 'data-request' status. QEMU and VirtualBox seemed to mark
the ready status anyway, but Hyper-V does not, resulting in the OS
freezing.