github: simplify actions into single action
Previously we had separate 'CI' and 'tag' workflows, one for regular
buils, and one for release builds that build extra utilities and push
artifacts. This wasn't optional because both workflows end up running
for every tag, and they're mostly duplicates.

To fix this, we're using `if' and checking for 'refs/tags/v' on every
step we only want run for version tags.