kp: allow non-constant log level
Previously, because we did a straight concat of the log level to the KP_STR
macros, a non constant log level did not work because it would just
result in a KP_STR macro that did not exist.

It's simple enough to replace the KP_STR with a %s, solving the problem,
but we want to avoid always doing that string replacement when in most
cases it is a constant. To make this work, we use __builtin_constant_p()
to check if the log level is constant, and use that to decide if we're
going to use %s or not.

The only other catch is that __builtin_constant_p() is not a macro, so
both code paths have to always be valid. Because we can't make all
KP_STR* macros valid, we instead move the quoting of the KP_STR outside
of it. This way, TP(KP_STR, level) produces an unquoted string, and even
if it is something like KP_STRfoobar (Which does not exist), it will
simply be quoted into "KP_STRfoobar", which is still a valid string to
be concatinated, making that code path always valid.

The use of __builtin_constant_p() means that only one code path will
ever be picked by the compiler, so we don't have to worry about the
other path being compiled or used when it is not needed or invalid.