mm: add user_check API for interfacting with userspace addresses
The current system of just assuming the address is always valid just
isn't correct. It was somewhat OK before we introduced memory mapping
and demand paging, but now the kernel can accidentally trigger userspace
page-faults at bad times and potentially screw everything up.

The solution to that is to segregate APIs that deal with userspace
pointers vs. kernel pointers (Which we already do, to an extent) and
introduce an API (the user_check API) for safely interacting with
userspace.

The design is based off of the Linux kernel, and is pretty simple. We
want page faults to result in a -EFAULT from the user_check API (That
way the kernel can then handle the error and ensure it does things like
free memory, or release locks).

To make this work, the page fault handler cannot kill the task, and also
cannot exit back to the same location. Thus, we manually set the address
to jump to on a page fault on a special field on the task. The
instruction pointer is then manually set to that location when a page
fault happens and the page fault handler exits.