spinlock: remove name and optional logging
This is a simple change, but touches a lot of the kernel. Spinlocks used
to have a name field, which held a string intended to indicate what the
spinlock was used for. This was used in conjunction with spinlock
logging for testing in the early days of the kernel.

With the current kernel, the spinlock logging is effectively unusable
since itn results in too much logging and too much of a slow down.
Consequently, the name field is also unused. Thus, we're removing the
name field and the spinlock logging option.

This change propgates down to a lot of our other locks. Mutex's and
wait-queues in paticular also had their names removed.