restructure Makefile build system
After restructing the project in the last commit, this fixes up the make
build system to work with the new structure, along with unifying a lot
of the locations by having the auxilary scripts be passed the locations
from the make files.

After this change, debugging is no longer started via running the script
directly, instead it is started via `make debug`, `make debug-ktest` and
`make debug-net`.