sys_sendto: pass a NULL sockaddr if length is zero
This was an accidental buf introduced in the user_buffer change.
Previously, the user pointer was passed directly up to the protocols in
regards to sendto's sockaddr buffer. This meant the protocols could
correctly check for NULL, and one or two do. Now however, we handle the
user_buffer for the sockaddr at the sys_sendto level and simply pass a
temporary buffer up to the protocol. Because the temporary buffer is
always valid, we are always passing it (Though we still pass the length
as is). The result is that the protocols that check `addr` see it as
always valid, which causes some issues (notably, TCP sendto does not
work).

The simple fix is to change the sys_sendto logic to send NULL for the
sockaddr buffer when the given length is zero.