kernel: x86: uninline hlt to avoid gdb bug
gdb appears to have a bug when a breakpoint happens in our hlt()
function. It's unclear why it happens, and it did not happen in our old
4.3 toolchain, only in our new 9.2 toolchain. The bug is aproximately
this:

internal-error: void inline_frame_this_id: Assertion !frame_id_eq(...

Because the kernel tends to spend msot of its time on the hlt() function
inside of the idle task, this means the majority of the time if we stop
execution in gdb to add breakpoints or etc. we hit this bug. To avoid
it, we're no longer using the hlt() inline function we wrote, and
instead adding the `asm volatile()` block directly.