Initial uapi split
The idea here is to separate the kernel headers into a "kernel-only" set
and a "userspace" set. When 'install-kernel-headers' is done, only the
"userspace" headers are copied, meaning userspace cannot unintentionally
include kernel-only stuff, and we don't need to have as much of the
ifdef stuff surounding kernel-only things.

One complicated part is what headers can include what. Somewhat
counterintuitively, just making a rule that "uapi headers only include
other uapi headers" isn't quite that simple. It creates an ugly
situation for the kernel headers, because you would have to manually
pull in all the kernel-only stuff along with the uapi stuff. It would
also require the `uapi` folder be present on the target system, since
the uapi headers would need to include things from the uapi folder.

Instead, the uapi headers only include things from the kernel headers,
and the kernel directory headers include their coresponding uapi header
with the same name. That way, when you include a header from the kernel
you automatically get all the uapi and all the kernel-only stuff, and
when the headers are installed on the user's system, only the stuff in
the uapi folder is copied, and the uapi path prefix is stripped away.
That way, when a uapi header includes something like "protura/types.h",
from the kernel point-of-view it pulls in the kernel-only
"protura/types.h", but on the user's system it pulls in the uapi
version, because that's the one present in "protura/types.h" on the
system.

The one issue with that approach is that if a uapi header is missing, it
is possible that the kernel will not be aware if the kernel-only version
does exist. To remedy this, a test ('check-uapi') is added that uses
`gcc` to verify the consistency when only using the uapi headers.

This part of the work wires up the uapi stuff by modifying
install-kernel-headers to copy from the uapi folder in regular 'include'
and the 'arch' folder. It also adds a very simple script that verifies
the uapi header set for consistency, which can be run with 'check-uapi'.
To make this initial code work, `protura/types.h` was also converted as
the first header.