docs: build.md: add information on buildtime dependencies
The current build, largely due to the external projects we build, has a
fair number of buildtime dependencies which are not documented (except
in the CI workflow). This updates the build.md file to include
information on all of the require buildtime dependencies, and simple
instructions for some distros (Just Ubuntu, at the moment).

Fixes: #24