block: redo block locking and waiting
Replaces the somewhat messy existing system for a much simpler one,
which is similar to the Linux Kernel's approach:

Instead of the existing wait-queue syncronous approach, when a block is
ready to be synced it is simply locked before being submitted. Then, the
block is kept locked until a driver like the IDE driver has finally
synced the block, setting the correct flags in the process and then
unlocking the block.

Thus, waiting for the block to be synced is simply a matter of locking
the block after submitting it. With that, drivers like the IDE driver no
longer need to be syncronous and can instead return after scheduling the
block without waiting for it to be written/read. The upper code can then
wait for the process to be done if they want.

Overall, this process is much simpler to manage with less moving parts,
and being able to queue off a bunch of blocks and then wait for them to
complete afterward is now possible. With that, the async design should
make the upper parts like inode cache much easier to make async as well.