waitpid: fix use-after-free of child
The child task is marked dead, but then we read from it to return
child-pid. The result is a use-after-free which can result in a
corrupted value being returned to userspace instead of the child's pid.

The fix is to save the child_pid to return before marking the child
dead.