ktest: Fix alignment of ktest section
Apparently, gcc likes to align struct's to 32 bytes. Previously we were
only aligning them to 8 bytes, which can result in the linker shoving
some padding into the ktest section and screwing the whole thing up.

Unfortunately the 32 byte alignment thing is very poorly documented, and
I'm not really sure why it happens. The closest I found to documentation
is a similar commit in the Linux kernel, where someone ran into this
same issue, and fixed it saying gcc's "struct alignment" in 4.5 was
changed to 32 bytes.