vfs: completely rewrite inode cache
This change completely rewrites the inode cache to better reflect the
inode cache in the Linux kernel. In paticular, inodes now have a variety
of states and state transitions representing what is currently happening
with the inode. The locking largely then protects some structures and
the state, allowing us to easily drop the lock in a lot of different
spots where we couldn't before and had to do a variety of convoluted
logic.

The locking itself can be a bit harder to follow, but is largely much
simpler as all if it is contained within the inode_table.c file, and no
callbacks are called when holding locks, making it much easier to avoid
deadlocks.