build: add kernel symbol resolution at runtime
This change includes a lot of build and code changes that add the
ability to resolve the name of the symbol containing a patciular
address. This is done via a fairly convoluted process of linking the
kernel multiple times, in this process:

1. First link contains no symbol table.
2. Second link contains the symbol table from the first pass
3. Third link (final link) contains the symbol table from the second
   link.

The symbol table can change between the first and second link due to the
size of the symbol table causing other symbols to move around. The
assumption is, however, that because no new symbols are introduced the
symbol table's from the first and second link should be the same size
(The symbol table's symbol is excluded to keep the table from getting
one entry larger).

There is then a helper function for looking up a symbol, and the symbols
are integrated into the backtrace and page fault logging. There is also
a simple test to verify the resulting symbol table does actually match
the one from the final kernel.