vfs: completely redo super-block and mount logic
The previous logic in sync.c was pretty bad, and likely had a fair
amount of race conditions or deadlocks in it. This new logic is similar
to the new inode_table logic and redoes all of the super-block and
mount/umount logic.

Super-blocks are now allocated before being initalized, similar to
inodes. The super-blocks themselves are then reference counted (to
ensure they don't get deallocated until they are completely unused) and
have a new 'umount_lock' that locks their new flags along with various
operations. The 'umount_lock' is exclusively used in the super-block
cache code and a separate lock is used in the ext2 code to protect the
block-groups and other such things.

The vfsmount entries are created in a similar fashion - they are created
up front and then the insertion is attempted automically, if it fails
then everything is rolled back.

umount is a paticularly touchy action, as we have to basically
determine if a umount can happen, and then perform it without allowing
any new actions to take place on the super-block. To achieve this the
inode_clear_super() functions was modified to take the root inode as an
argument, and then automatically drops the reference to it if the root
inode is the only one still existing. Then, the super-block logic holds
the umount_lock while calling inode_clear_super(), and the vfsmount is
marked with the VFS_MOUNT_UNMOUNTING flag, which will make anybody
attempting to use that mountpoint wait on the umount_queue. The queue is
then triggered when either the mountpoint is gone, or the umount failed.

Sync is also a similar situation (but simpler). We loop over the
super-block list, grab a reference to each super-block, and then take
the umount_lock and check the flags to figure out if the super-block can
be synced. If it can, we sync the inodes and super-block while holding
the umount_lock. After all the super-blocks are synced, we sync all the
block-devices.