mm: panic() on recursive page-fault
A recursive page-fault indicates some kind of problem - we should not
produce a page fault when handling an existing page-fault. This can
happen in weird memory corruption situations and it is useful to catch
them and panic early, as otherwise it would simply go on forever,
overflowing the stack and eventually triplefaulting.